+++ .github/workflows/data-processing-pipeline.yml
@@ -1,10 +1,6 @@
 name: Data Processing Pipeline (Prefect Flow)
 
-on: 
-  push:
-    branches:
-      - main
-  workflow_dispatch:
+on: 'push'
 
 jobs:
   deploy:
